Move SIP module stuff to Msf::Exploit::Remote::SIP
