+++ lib/msf/core/exploit/mixins.rb
@@ -57,6 +57,7 @@ require 'msf/core/exploit/wdbrpc'
 require 'msf/core/exploit/wdbrpc_client'
 require 'msf/core/exploit/afp'
 require 'msf/core/exploit/realport'
+require 'msf/core/exploit/sip'
 
 # Telephony
 require 'msf/core/exploit/dialup'
