+++ lib/rex/proto/sip.rb
@@ -2,4 +2,3 @@
 
 # SIP protocol support
 require 'rex/proto/sip/response'
-require 'rex/proto/sip/util'
